Couple Device schema with inventory
